Fix update of folder templates upon start of workout
When starting a new workout from a template, the template becomes the workout and a copy of the template should be made to restore the template in its source folder. A copy was being made but this was not reflected in its source folder and the completed workout remained in the folders. This fix ensures that the source folder is updated to retain the template and remove the workout. To enable the view to update when this happens, the folder view has to be separated into it's own view to observe it's own folder.